Add files via upload
Well, here is the 1st part of my pattern recognition project, for now I just implemented a way for the algorithm to decide the number of colors that it will use to simplify a picture and also simplify it in white/black spectrum. That way we will reduce the noise. Now I am working on a way for the algorythm to draw the convex hull of each cluster of the image. I would like to reduct the noise even more after doing that. Will figure it out soon enough i guess.